requirements.txt
pandas>=2.2.0
numpy>=1.26.0
python-igraph>=0.11.0
streamlit>=1.35.0
pyarrow>=16.0.0
rapidfuzz>=3.9.0
pytest>=8.2.0
